% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{say_hello}
\alias{say_hello}
\title{Say Hello Function}
\usage{
say_hello()
}
\value{
None
}
\description{
This function prints "Hello, world!" to the console.
}
\details{
This function simply prints a greeting message to the console.
}
\examples{
\dontrun{
if(interactive()){
  say_hello()
  }
}

}
