% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{sum_two_numbers}
\alias{sum_two_numbers}
\title{Sum Two Numbers Function}
\usage{
sum_two_numbers(a, b)
}
\arguments{
\item{a}{The first number.}

\item{b}{The second number.}
}
\value{
The sum of the two numbers.
}
\description{
This function calculates the sum of two numbers.
}
\details{
This function uses the `sum` function from the `dplyr` package to compute the sum of two numbers.
}
\examples{
\dontrun{
if(interactive()){
  result <- sum_two_numbers(5, 3)
  print(result)  # Output: 8
  }
}


}
